- Organise solution files into two main solutions - nature-based solution and engineering solution
